+++ wxPython/build/Mac/setup_daemon.py
@@ -1,7 +1,11 @@
 from setuptools import setup
 import os
 
-version = open('/Users/yanone/Code/py/git/typeworld/guiapp/currentVersion.txt', 'r').read().strip()
+from ynlib.web import GetHTTP
+version = GetHTTP('https://api.type.world/latestUnpublishedVersion/world.type.guiapp/mac/')
+if version == 'n/a':
+    print('Can’t get version number')
+    sys.exit(1)
 
 os.system('rm -rf ~/Code/TypeWorldApp/apps/Mac/Type.World.%s.app' % version)
